// src/app/ejercicios/dragonball-api/interfaces/personaje-detalle.ts
export interface PersonajeDetalle {
  id: number;
  nombre: string;
  imagen: string;
  raza: string;
  genero: string;
  ki: number;
  maxKi:number;
  descripcion:string;
  afiliacion: string;
  planetaOrigen:string;
  IdPlanetaOrigen:number;
  planetaEsDestruido:boolean;
  transformaciones:number;
}